requirements.txt
# V-JEPA 2 CVS Classification Dependencies
# Install PyTorch first: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

transformers>=4.40.0
accelerate>=0.30.0
timm>=0.9.0
einops>=0.7.0
pillow>=10.0.0
tqdm>=4.66.0
scikit-learn>=1.4.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
